requirements.txt
flask>=3.0,<4.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
composio-core>=0.7.0
